move view_supported_devices to device [pr]
