+++ test/test_subbuffer.py
@@ -1,7 +1,6 @@
 import unittest
 from tinygrad import Device, dtypes, Tensor
-from tinygrad.device import Buffer
-from tinygrad.engine.lazy import view_supported_devices
+from tinygrad.device import Buffer, view_supported_devices
 
 @unittest.skipIf(Device.DEFAULT not in view_supported_devices, "subbuffer not supported")
 class TestSubBuffer(unittest.TestCase):
